Extractig common plugin conventions accross Android modules
